// src/app/store/shared/enum/test-executeActionTypes.enum.ts
export enum TestExecuteActionTypes {
    NextQues = 'INC_QUES' ,
    PrevQues = 'DEC_QUES' ,
    SkipQues = 'SKIP_QUES' ,

    NextFilter = 'NEXT_FILTER' ,
    PrevFilter = 'PREVFILTER' ,

    SubmitTest ='SUBMIT_TEST' ,

    SubmitAns ='SUBMIT_ANS' ,

    MarkQues ='MARK_QUES' ,

    IntialFetchQuestion = 'INTIAL_FETCH_QUESTION',
    SectionQuesFilter = "SECTION_QUES_FILTER",

}